--- __mocks__/firebase-admin.js
@@ -1,12 +1,0 @@
-const firebaseAdminMock = jest.createMockFromModule('firebase-admin')
-
-const mockAuthValue = {
-  createCustomToken: jest.fn(() => Promise.resolve(null)),
-  verifyIdToken: jest.fn(() => Promise.resolve(null)),
-  getUser: jest.fn(() => Promise.resolve(null)),
-}
-firebaseAdminMock.auth = jest.fn(() => mockAuthValue)
-
-firebaseAdminMock.apps = []
-
-module.exports = firebaseAdminMock
